backend/requirements.txt
fastapi
uvicorn[standard]
# motor
# beanie
# pydantic
# python-dotenv
# python-jose[cryptography]
# passlib[bcrypt]
# python-multipart
# python -m pip install "pymongo[srv]"
motor
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic[email]